refactor: file-private preview fixture 를 중앙 집결지로 통합 (preview-policy.md §6)
전수조사 갈래 A — 자동 수정 결과:

1. feature/diary/diarywrite/DiaryPreviewData.kt 의 fixture 들을
   domain/model/preview/DiaryPreviewData.kt 로 통합 (sampleWriteUiState 등).
   feature/ 안의 별도 파일 삭제. 사용처 import 갱신.
2. lecturedetail/currenttable/LectureColorSelectorScreen.kt 의 file-private
   previewBuiltInPickerColors alias 제거. 사용처에서 LectureColor.Custom.Default
   를 직접 인라인 (도메인 default 가 이미 정의되어 있어 별도 fixture 가치 없음).

login/themeconfig/search/settings 디렉토리는 file-private fixture 패턴 0건
확인 — 갈래 A 자동 수정 대상 없음.
